+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2019-01-02
 ### Added
 - The Exclusive method, that returns a Tuple of two Ranges that represent the parts they do not have in common, including the ^ operator
 - A RangeCollection<T>, that holds a list of Ranges
@@ -56,7 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This CHANGELOG file to hopefully serve as an evolving example of a
   standardized open source project CHANGELOG.
 
-[Unreleased]: https://github.com/reynj/reynj/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/reynj/reynj/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/reynj/reynj/compare/v0.1.5...v0.2.0
 [0.1.5]: https://github.com/reynj/reynj/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/reynj/reynj/compare/v0.1.2...v0.1.4
 [0.1.3]: https://github.com/reynj/reynj/compare/v0.1.2...v0.1.3